# api/openapi/internal.yaml
openapi: 3.0.3
info:
  title: User Management Internal API
  version: 1.0.0
  description: |
    Internal REST surface for full user CRUD. Do not expose `/internal` to the public internet.
    Service-to-service callers that prefer HTTP should use this API. Prefer gRPC (`usermgmt.v1.UserService`) on the mesh when possible.
servers:
  - url: http://localhost:8081
    description: Local internal HTTP listener
paths:
  /internal/v1/users:
    get:
      summary: List users
      parameters:
        - in: query
          name: page_size
          schema:
            type: integer
            default: 20
            maximum: 100
        - in: query
          name: page_token
          schema:
            type: string
      responses:
        "200":
          description: A page of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
    post:
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserWrite"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid payload
        "409":
          description: Username or email already exists
  /internal/v1/users/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get a user
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    put:
      summary: Replace a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserWrite"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "409":
          description: Username or email already exists
    patch:
      summary: Partially update a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      summary: Delete a user
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
components:
  schemas:
    User:
      type: object
      required: [id, username, email, params, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        params:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserWrite:
      type: object
      required: [username, email]
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        params:
          type: object
          additionalProperties: true
    UserPatch:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        params:
          type: object
          additionalProperties: true
    UserList:
      type: object
      required: [users]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        next_page_token:
          type: string